refactor and create typeDetectionService and readerService
